requirements.txt
fastapi==0.115.8
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.4
python-dotenv==1.0.1
python-multipart==0.0.20
streamlit==1.44.1
typeguard==4.3.0
uvicorn==0.34.0
